organize pod/recursion
